Verified transforms are being passed to the shader
This just moves the triangle over a bit in the X direction. That the
triangle actually moved confirms the vertex multiplication in the vertex
shader is working.

This is a step towards getting nalgebra able to move things they way I
want.

I think I may soon outgrow the Triangle object.